Correctly separate the file
