+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.8",
   "description": "Porting of the original Magium app to JS",
   "license": "MIT",
-  "main": "main.js",
+  "main": "main_electron.js",
   "scripts": {
     "start:server": "nodemon main_node.js",
     "start:electron": "electron .",
